backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
python-multipart==0.0.20
python-dotenv==1.0.1
google-generativeai==0.8.3
faiss-cpu==1.13.2
numpy==1.26.4
tqdm==4.66.5
sentence-transformers==3.3.1
